fix: rewardPercent

+++ moysklad/commission_report_in.go
@@ -39,7 +39,7 @@ type CommissionReportIn struct {
 	Published                     *bool                                        `json:"published,omitempty"`
 	Rate                          *Rate                                        `json:"rate,omitempty"`
 	ReturnToCommissionerPositions *Positions[CommissionReportInReturnPosition] `json:"returnToCommissionerPositions,omitempty"`
-	RewardPercent                 *int                                         `json:"rewardPercent,omitempty"`
+	RewardPercent                 *float64                                     `json:"rewardPercent,omitempty"`
 	Payments                      *Payments                                    `json:"payments,omitempty"`
 	SalesChannel                  *SalesChannel                                `json:"salesChannel,omitempty"`
 	Shared                        *bool                                        `json:"shared,omitempty"`
